requirements.txt
gTTS==2.5.4
python-telegram-bot==21.10

